refactor(config): split god file by responsibility
config.go (1659 行, 6 职责混杂) 按职责拆分为 5 个文件:
- config_types.go:    struct/const 定义 + 数据方法
- config_defaults.go: Default() + propagate/normalize 辅助
- config_loader.go:   Load/loadRecursive + 路径解析
- config_env.go:      env-var reflect 映射
- config.go:          共享工具 (ExpandEnv/Validate/HotplexHome)

纯文件重组, 零逻辑变更. build/vet/test 全通过.

Refs #338
