+++ internal/config/config.go
@@ -1,20 +1,21 @@
+// Package config defines the gateway configuration types, default values,
+// loading pipeline, and environment-variable overrides.
+//
+// The configuration system is split across several files by responsibility:
+//   - config_types.go    — struct and constant definitions
+//   - config_defaults.go — Default() and propagation/normalization helpers
+//   - config_loader.go   — Load(), loadRecursive(), path resolution
+//   - config_env.go      — environment-variable mapping via reflection
+//   - config.go          — shared utilities (env expansion, validation, paths)
 package config
 
 import (
-	"errors"
-	"fmt"
 	"log/slog"
 	"os"
 	"path/filepath"
-	"reflect"
 	"regexp"
-	"slices"
-	"strconv"
 	"strings"
 	"sync"
-	"time"
-
-	"github.com/spf13/viper"
 )
 
 var envVarRe = regexp.MustCompile(`\$\{([^}:]+)(?::-([^}]*))?\}`)
@@ -102,1272 +103,6 @@ func (c *Config) Validate() []string {
 	return errs
 }
 
-// ─── Config structs ───────────────────────────────────────────────────────────
-
-// Config holds all gateway configuration.
-type Config struct {
-	Gateway     GatewayConfig   `mapstructure:"gateway"`
-	DB          DBConfig        `mapstructure:"db"`
-	Worker      WorkerConfig    `mapstructure:"worker"`
-	Security    SecurityConfig  `mapstructure:"security"`
-	Session     SessionConfig   `mapstructure:"session"`
-	Pool        PoolConfig      `mapstructure:"pool"`
-	Log         LogConfig       `mapstructure:"log"`
-	Admin       AdminConfig     `mapstructure:"admin"`
-	WebChat     WebChatConfig   `mapstructure:"webchat"`
-	Messaging   MessagingConfig `mapstructure:"messaging"`
-	AgentConfig AgentConfig     `mapstructure:"agent_config"`
-	Skills      SkillsConfig    `mapstructure:"skills"`
-	Cron        CronConfig      `mapstructure:"cron"`
-	Webhook     WebhookConfig   `mapstructure:"webhook"`
-	Events      EventsConfig    `mapstructure:"events"`
-	Inherits    string          `mapstructure:"inherits"` // path to parent config file; "" = no inheritance
-
-	// ResolvedAPIKeyUsers is the runtime map of expanded API key value → userID.
-	// Populated by resolveAPIKeyUsers() during load. Nil when no mapping configured.
-	ResolvedAPIKeyUsers map[string]string `mapstructure:"-"`
-}
-
-// MessagingConfig holds messaging platform adapter settings.
-// Shared defaults (WorkerType, STTConfig, TTSConfig) are set at this level and propagated
-// to each platform config via propagateMessagingDefaults().
-// Access control fields (DMPolicy, GroupPolicy, RequireMention, AllowFrom, AllowDMFrom, AllowGroupFrom) support per-bot overrides with platform-level fallback.
-// Priority: platform-level > messaging-level > Default().
-type MessagingConfig struct {
-	TurnSummaryEnabled bool `mapstructure:"turn_summary_enabled"`
-
-	// Shared defaults — platforms inherit; platform-level overrides take precedence.
-	WorkerType string `mapstructure:"worker_type"`
-	STTConfig  `mapstructure:",squash"`
-	TTSConfig  `mapstructure:",squash"`
-
-	// Platform-specific configs.
-	Slack   SlackConfig   `mapstructure:"slack"`
-	Feishu  FeishuConfig  `mapstructure:"feishu"`
-	Yuanxin YuanxinConfig `mapstructure:"yuanxin"`
-}
-
-// STT constants for provider values.
-const (
-	STTProviderLocal       = "local"
-	STTProviderFeishu      = "feishu"
-	STTProviderFeishuLocal = "feishu+local"
-)
-
-// STTConfig holds speech-to-text configuration shared across messaging adapters.
-type STTConfig struct {
-	// Provider: "local" (external command), "feishu" (cloud API),
-	// "feishu+local" (cloud primary, local fallback), "" (disabled).
-	Provider string `mapstructure:"stt_provider"`
-	// LocalCmd is the command template. {file} is replaced with the audio file path.
-	// If {file} is present → ephemeral mode (fork per request).
-	// If absent → persistent mode (long-lived subprocess, stdin/stdout JSON protocol).
-	LocalCmd string `mapstructure:"stt_local_cmd"`
-	// LocalIdleTTL controls auto-shutdown of persistent subprocess. 0 = disabled.
-	LocalIdleTTL time.Duration `mapstructure:"stt_local_idle_ttl"`
-}
-
-// IsPersistent returns true when LocalCmd uses persistent mode (no {file} placeholder).
-func (c STTConfig) IsPersistent() bool {
-	return c.LocalCmd != "" && !strings.Contains(c.LocalCmd, "{file}")
-}
-
-// FillFrom propagates zero-value fields from defaults into c.
-func (c *STTConfig) FillFrom(defaults STTConfig) {
-	if c.Provider == "" {
-		c.Provider = defaults.Provider
-	}
-	if c.LocalCmd == "" {
-		c.LocalCmd = defaults.LocalCmd
-	}
-	if c.LocalIdleTTL == 0 {
-		c.LocalIdleTTL = defaults.LocalIdleTTL
-	}
-}
-
-// TTSConfig holds text-to-speech settings. Squashed into platform configs.
-type TTSConfig struct {
-	// TTSEnabled controls whether voice replies are sent for voice-triggered turns.
-	TTSEnabled bool `mapstructure:"tts_enabled"`
-	// Provider: "edge" (Microsoft Edge TTS, free), "edge+moss" (Edge primary + MOSS-TTS-Nano CPU fallback), "" (disabled).
-	TTSProvider string `mapstructure:"tts_provider"`
-	// Voice name for Edge TTS (e.g. "zh-CN-XiaoxiaoNeural", "zh-CN-YunxiNeural").
-	Voice string `mapstructure:"tts_voice"`
-	// MaxChars limits the LLM summary length before TTS synthesis.
-	MaxChars int `mapstructure:"tts_max_chars"`
-	// MossModelDir is the directory containing MOSS-TTS-Nano ONNX model assets (used when provider is "edge+moss").
-	MossModelDir string `mapstructure:"tts_moss_model_dir"`
-	// MossVoice is the MOSS built-in voice preset name (default "Xiaoyu").
-	MossVoice string `mapstructure:"tts_moss_voice"`
-	// MossPort is the localhost port for the MOSS sidecar HTTP server (default 18083).
-	MossPort int `mapstructure:"tts_moss_port"`
-	// MossIdleTimeout controls how long the MOSS sidecar stays resident after its
-	// last use before being automatically shut down. Default 30m.
-	MossIdleTimeout time.Duration `mapstructure:"tts_moss_idle_timeout"`
-	// MossCpuThreads controls ONNX Runtime intra-op threads for the MOSS sidecar (0 = auto-detect = physical cores).
-	MossCpuThreads int `mapstructure:"tts_moss_cpu_threads"`
-}
-
-// FillFrom propagates zero-value fields from defaults into c.
-func (c *TTSConfig) FillFrom(defaults TTSConfig) {
-	if !c.TTSEnabled && defaults.TTSEnabled {
-		c.TTSEnabled = defaults.TTSEnabled
-	}
-	if c.TTSProvider == "" {
-		c.TTSProvider = defaults.TTSProvider
-	}
-	if c.Voice == "" {
-		c.Voice = defaults.Voice
-	}
-	if c.MaxChars == 0 {
-		c.MaxChars = defaults.MaxChars
-	}
-	if c.MossModelDir == "" {
-		c.MossModelDir = defaults.MossModelDir
-	}
-	if c.MossVoice == "" {
-		c.MossVoice = defaults.MossVoice
-	}
-	if c.MossPort == 0 {
-		c.MossPort = defaults.MossPort
-	}
-	if c.MossIdleTimeout == 0 {
-		c.MossIdleTimeout = defaults.MossIdleTimeout
-	}
-	if c.MossCpuThreads == 0 {
-		c.MossCpuThreads = defaults.MossCpuThreads
-	}
-}
-
-// MessagingPlatformConfig holds settings shared by all messaging adapters (Slack, Feishu, etc.).
-type MessagingPlatformConfig struct {
-	Enabled        bool     `mapstructure:"enabled"`
-	WorkerType     string   `mapstructure:"worker_type"`
-	WorkDir        string   `mapstructure:"work_dir"`
-	DMPolicy       string   `mapstructure:"dm_policy"`
-	GroupPolicy    string   `mapstructure:"group_policy"`
-	RequireMention bool     `mapstructure:"require_mention"`
-	AllowFrom      []string `mapstructure:"allow_from"`
-	AllowDMFrom    []string `mapstructure:"allow_dm_from"`
-	AllowGroupFrom []string `mapstructure:"allow_group_from"`
-
-	InjectExclude []string `mapstructure:"inject_exclude"` // platform-level default: files to skip from agent config injection
-	STTConfig     `mapstructure:",squash"`
-	TTSConfig     `mapstructure:",squash"`
-}
-
-// MaxBotsPerPlatform is the maximum number of bots allowed per messaging platform.
-const MaxBotsPerPlatform = 10
-
-// SlackConfig holds Slack Socket Mode adapter settings.
-// Supports single-bot (top-level bot_token/app_token) and multi-bot (bots[]) modes.
-// normalizeSlackBots() resolves the two into a unified Bots slice.
-type SlackConfig struct {
-	MessagingPlatformConfig `mapstructure:",squash"`
-
-	// Single-bot credentials (backward compatible).
-	BotToken string `mapstructure:"bot_token"`
-	AppToken string `mapstructure:"app_token"`
-
-	SocketMode          bool          `mapstructure:"socket_mode"`
-	AssistantAPIEnabled *bool         `mapstructure:"assistant_api_enabled"`
-	ReconnectBaseDelay  time.Duration `mapstructure:"reconnect_base_delay"`
-	ReconnectMaxDelay   time.Duration `mapstructure:"reconnect_max_delay"`
-
-	// Branding for Assistant status (paid workspaces).
-	DisplayName string `mapstructure:"display_name,omitempty"`
-	IconEmoji   string `mapstructure:"icon_emoji,omitempty"`
-
-	// Multi-bot configuration. When non-empty, takes precedence over top-level credentials.
-	Bots []SlackBotConfig `mapstructure:"bots"`
-}
-
-// SlackBotConfig holds credentials and per-bot overrides for a single Slack bot.
-type SlackBotConfig struct {
-	Name       string `mapstructure:"name"`
-	BotToken   string `mapstructure:"bot_token"`
-	AppToken   string `mapstructure:"app_token"`
-	WorkerType string `mapstructure:"worker_type,omitempty"`
-	WorkDir    string `mapstructure:"work_dir,omitempty"`
-	Sandbox    string `mapstructure:"sandbox,omitempty"`
-	ACPCommand string `mapstructure:"acp_command,omitempty"` // per-bot ACP agent binary override
-
-	// Per-bot access control (falls back to platform-level when empty).
-	DMPolicy       string   `mapstructure:"dm_policy,omitempty"`
-	GroupPolicy    string   `mapstructure:"group_policy,omitempty"`
-	RequireMention *bool    `mapstructure:"require_mention,omitempty"`
-	AllowFrom      []string `mapstructure:"allow_from,omitempty"`
-	AllowDMFrom    []string `mapstructure:"allow_dm_from,omitempty"`
-	AllowGroupFrom []string `mapstructure:"allow_group_from,omitempty"`
-
-	// Per-bot agent config injection override (falls back to platform-level when nil).
-	// No omitempty: inject_exclude: [] must produce a non-nil empty slice to
-	// explicitly clear the parent-level exclusion (nil = inherit, [] = clear).
-	InjectExclude []string `mapstructure:"inject_exclude"`
-
-	// Per-bot branding override (falls back to platform-level when empty).
-	DisplayName string `mapstructure:"display_name,omitempty"`
-	IconEmoji   string `mapstructure:"icon_emoji,omitempty"`
-
-	STTConfig `mapstructure:",squash"`
-	TTSConfig `mapstructure:",squash"`
-}
-
-// FeishuConfig holds Feishu WebSocket adapter settings.
-// Supports single-bot (top-level app_id/app_secret) and multi-bot (bots[]) modes.
-// normalizeFeishuBots() resolves the two into a unified Bots slice.
-type FeishuConfig struct {
-	MessagingPlatformConfig `mapstructure:",squash"`
-
-	// Single-bot credentials (backward compatible).
-	AppID     string `mapstructure:"app_id"`
-	AppSecret string `mapstructure:"app_secret"`
-
-	// Multi-bot configuration. When non-empty, takes precedence over top-level credentials.
-	Bots []FeishuBotConfig `mapstructure:"bots"`
-}
-
-// FeishuBotConfig holds credentials and per-bot overrides for a single Feishu bot.
-type FeishuBotConfig struct {
-	Name       string `mapstructure:"name"`
-	AppID      string `mapstructure:"app_id"`
-	AppSecret  string `mapstructure:"app_secret"`
-	WorkerType string `mapstructure:"worker_type,omitempty"`
-	WorkDir    string `mapstructure:"work_dir,omitempty"`
-	Sandbox    string `mapstructure:"sandbox,omitempty"`
-	ACPCommand string `mapstructure:"acp_command,omitempty"` // per-bot ACP agent binary override
-
-	// Per-bot access control (falls back to platform-level when empty).
-	DMPolicy       string   `mapstructure:"dm_policy,omitempty"`
-	GroupPolicy    string   `mapstructure:"group_policy,omitempty"`
-	RequireMention *bool    `mapstructure:"require_mention,omitempty"`
-	AllowFrom      []string `mapstructure:"allow_from,omitempty"`
-	AllowDMFrom    []string `mapstructure:"allow_dm_from,omitempty"`
-	AllowGroupFrom []string `mapstructure:"allow_group_from,omitempty"`
-
-	// Per-bot agent config injection override (falls back to platform-level when nil).
-	// No omitempty: inject_exclude: [] must produce a non-nil empty slice to
-	// explicitly clear the parent-level exclusion (nil = inherit, [] = clear).
-	InjectExclude []string `mapstructure:"inject_exclude"`
-
-	STTConfig `mapstructure:",squash"`
-	TTSConfig `mapstructure:",squash"`
-}
-
-// YuanxinConfig holds Yuanxin Pulsar adapter settings.
-type YuanxinConfig struct {
-	MessagingPlatformConfig `mapstructure:",squash"`
-
-	Tenant        string `mapstructure:"tenant"`
-	Namespace     string `mapstructure:"namespace"`
-	PulsarURL     string `mapstructure:"pulsar_url"`
-	AppID         string `mapstructure:"app_id"`
-	ProducerTopic string `mapstructure:"producer_topic"`
-}
-
-type AdminConfig struct {
-	Enabled            bool                `mapstructure:"enabled"`
-	Addr               string              `mapstructure:"addr"`
-	Tokens             []string            `mapstructure:"tokens"`
-	TokenScopes        map[string][]string `mapstructure:"token_scopes"`
-	DefaultScopes      []string            `mapstructure:"default_scopes"`
-	IPWhitelistEnabled bool                `mapstructure:"ip_whitelist_enabled"`
-	AllowedCIDRs       []string            `mapstructure:"allowed_cidrs"`
-	RateLimitEnabled   bool                `mapstructure:"rate_limit_enabled"`
-	RequestsPerSec     int                 `mapstructure:"requests_per_sec"`
-	Burst              int                 `mapstructure:"burst"`
-}
-
-// LogConfig holds logging settings.
-type LogConfig struct {
-	Level  string `mapstructure:"level"`
-	Format string `mapstructure:"format"` // "json" or "text"
-}
-
-// WebChatConfig holds webchat UI serving settings.
-type WebChatConfig struct {
-	Addr    string `mapstructure:"addr"`    // informational: banner display
-	Enabled bool   `mapstructure:"enabled"` // serve embedded webchat SPA from gateway
-}
-
-// GatewayConfig holds WebSocket gateway settings.
-type GatewayConfig struct {
-	Addr               string        `mapstructure:"addr"`
-	ReadBufferSize     int           `mapstructure:"read_buffer_size"`
-	WriteBufferSize    int           `mapstructure:"write_buffer_size"`
-	PingInterval       time.Duration `mapstructure:"ping_interval"`
-	PongTimeout        time.Duration `mapstructure:"pong_timeout"`
-	WriteTimeout       time.Duration `mapstructure:"write_timeout"`
-	IdleTimeout        time.Duration `mapstructure:"idle_timeout"`
-	MaxFrameSize       int64         `mapstructure:"max_frame_size"`
-	BroadcastQueueSize int           `mapstructure:"broadcast_queue_size"`
-
-	// PlatformWriteBuffer is the per-conn channel capacity for async platform writes.
-	// 64 slots accommodate ~8 batches at 120ms coalesce window, providing ample headroom
-	// even under burst conditions without excessive memory overhead.
-	PlatformWriteBuffer int `mapstructure:"platform_write_buffer"`
-	// PlatformDropThreshold is the fill level at which droppable events (delta/raw)
-	// begin being silently dropped. Set to 87.5% of PlatformWriteBuffer to provide
-	// backpressure relief while preserving space for guaranteed events.
-	PlatformDropThreshold int `mapstructure:"platform_drop_threshold"`
-	// DeltaCoalesceInterval is the time window for batching consecutive delta events.
-	// 120ms targets Feishu CardKit's 10 updates/sec per-card rate limit (8.3/sec with margin),
-	// while keeping first-token latency well under the 200ms human perception threshold.
-	// At 100 tok/s input, this yields ~12x API call reduction.
-	DeltaCoalesceInterval time.Duration `mapstructure:"delta_coalesce_interval"`
-	// DeltaCoalesceSize is the rune threshold for immediate delta flush, serving as a
-	// burst safety valve. 200 runes ≈ 40 tokens triggers early flush only during spikes,
-	// while average batches at 100 tok/s / 120ms ≈ 48 runes stay well below this threshold.
-	DeltaCoalesceSize int `mapstructure:"delta_coalesce_size"`
-}
-
-// DBConfig holds database settings.
-// Supports both SQLite and PostgreSQL backends.
-// For backward compatibility, legacy flat fields (Path, WALMode, etc.) are kept
-// on DBConfig. New code should prefer the structured SQLite.* or Postgres.* fields.
-type DBConfig struct {
-	// Driver specifies the database driver: "sqlite" (default) or "postgres".
-	// When set to "postgres", the Postgres sub-config is used instead of SQLite.
-	Driver string `mapstructure:"driver"`
-
-	// SQLite-specific configuration. Also the target for legacy flat fields
-	// when using the sqlite driver (default).
-	SQLite SQLiteConfig `mapstructure:"sqlite"`
-
-	// Postgres-specific configuration. Active when Driver is "postgres".
-	Postgres PostgresConfig `mapstructure:"postgres"`
-
-	// ── Legacy flat fields (deprecated, use SQLite.* instead) ──
-	Path              string        `mapstructure:"path"`
-	EventsPath        string        `mapstructure:"events_path"` // Deprecated: events table now lives in hotplex.db (same as Path)
-	WALMode           bool          `mapstructure:"wal_mode"`
-	BusyTimeout       time.Duration `mapstructure:"busy_timeout"`
-	MaxOpenConns      int           `mapstructure:"max_open_conns"`
-	VacuumThreshold   float64       `mapstructure:"vacuum_threshold"`
-	CacheSizeKiB      int           `mapstructure:"cache_size_kib"`
-	MmapSizeMiB       int           `mapstructure:"mmap_size_mib"`
-	WalAutoCheckpoint int           `mapstructure:"wal_autocheckpoint"`
-}
-
-// DSN returns the connection string for the configured database driver.
-// Delegates to the active sub-config's DSN method based on Driver.
-func (d DBConfig) DSN() string {
-	switch strings.ToLower(d.Driver) {
-	case "postgres", "pg", "postgresql":
-		return d.Postgres.DSN()
-	default:
-		return d.SQLite.DSN()
-	}
-}
-
-// EffectiveSQLitePath returns the effective SQLite database path,
-// preferring the structured SQLite.Path with legacy flat Path as fallback.
-func (d DBConfig) EffectiveSQLitePath() string {
-	if d.SQLite.Path != "" {
-		return d.SQLite.Path
-	}
-	return d.Path
-}
-
-// EffectiveMaxOpenConns returns the effective max open connections,
-// preferring the structured SQLite.MaxOpenConns with legacy flat MaxOpenConns as fallback.
-func (d DBConfig) EffectiveMaxOpenConns() int {
-	if d.SQLite.MaxOpenConns > 0 {
-		return d.SQLite.MaxOpenConns
-	}
-	return d.MaxOpenConns
-}
-
-// EffectiveWALMode returns the effective WAL mode setting.
-// Note: explicitly setting WALMode=false (zero value) falls through to the legacy field.
-// This is acceptable because WALMode=false is not a meaningful config — users who want
-// to disable WAL should omit the field and set the legacy wal_mode: false instead.
-func (d DBConfig) EffectiveWALMode() bool {
-	if d.SQLite.WALMode {
-		return true
-	}
-	return d.WALMode
-}
-
-// EffectiveBusyTimeout returns the effective busy timeout for SQLite.
-func (d DBConfig) EffectiveBusyTimeout() time.Duration {
-	if d.SQLite.BusyTimeout > 0 {
-		return d.SQLite.BusyTimeout
-	}
-	return d.BusyTimeout
-}
-
-// EffectiveCacheSizeKiB returns the effective SQLite cache size in KiB.
-func (d DBConfig) EffectiveCacheSizeKiB() int {
-	if d.SQLite.CacheSizeKiB > 0 {
-		return d.SQLite.CacheSizeKiB
-	}
-	return d.CacheSizeKiB
-}
-
-// EffectiveMmapSizeMiB returns the effective SQLite mmap size in MiB.
-func (d DBConfig) EffectiveMmapSizeMiB() int {
-	if d.SQLite.MmapSizeMiB > 0 {
-		return d.SQLite.MmapSizeMiB
-	}
-	return d.MmapSizeMiB
-}
-
-// EffectiveWalAutoCheckpoint returns the effective WAL auto-checkpoint threshold.
-func (d DBConfig) EffectiveWalAutoCheckpoint() int {
-	if d.SQLite.WalAutoCheckpoint > 0 {
-		return d.SQLite.WalAutoCheckpoint
-	}
-	return d.WalAutoCheckpoint
-}
-
-// SQLiteConfig holds SQLite-specific database settings.
-type SQLiteConfig struct {
-	Path              string        `mapstructure:"path"`
-	WALMode           bool          `mapstructure:"wal_mode"`
-	BusyTimeout       time.Duration `mapstructure:"busy_timeout"`
-	MaxOpenConns      int           `mapstructure:"max_open_conns"`
-	VacuumThreshold   float64       `mapstructure:"vacuum_threshold"`
-	CacheSizeKiB      int           `mapstructure:"cache_size_kib"`
-	MmapSizeMiB       int           `mapstructure:"mmap_size_mib"`
-	WalAutoCheckpoint int           `mapstructure:"wal_autocheckpoint"`
-}
-
-// DSN returns the SQLite database path. Defaults to ":memory:" when empty.
-func (s SQLiteConfig) DSN() string {
-	if s.Path == "" {
-		return ":memory:"
-	}
-	return s.Path
-}
-
-// PostgresConfig holds PostgreSQL-specific database settings.
-type PostgresConfig struct {
-	ConnStr      string `mapstructure:"dsn"`
-	MaxOpenConns int    `mapstructure:"max_open_conns"`
-}
-
-// DSN returns the PostgreSQL connection string. Returns empty when ConnStr is not configured.
-func (p PostgresConfig) DSN() string {
-	return p.ConnStr
-}
-
-// WorkerConfig holds per-worker defaults.
-type WorkerConfig struct {
-	MaxLifetime      time.Duration        `mapstructure:"max_lifetime"`
-	IdleTimeout      time.Duration        `mapstructure:"idle_timeout"`
-	ExecutionTimeout time.Duration        `mapstructure:"execution_timeout"`
-	TurnTimeout      time.Duration        `mapstructure:"turn_timeout"`
-	EnvBlocklist     []string             `mapstructure:"env_blocklist"`
-	DefaultWorkDir   string               `mapstructure:"default_work_dir"`
-	PIDDir           string               `mapstructure:"pid_dir"`
-	AutoRetry        AutoRetryConfig      `mapstructure:"auto_retry"`
-	OpenCodeServer   OpenCodeServerConfig `mapstructure:"opencode_server"`
-	ClaudeCode       ClaudeCodeConfig     `mapstructure:"claude_code"`
-	CodexCLI         CodexCLIConfig       `mapstructure:"codex_cli"`
-	ACP              ACPConfig            `mapstructure:"acp"`
-	Environment      []string             `mapstructure:"environment"`
-}
-
-// MCPServerConfig defines a single MCP server for worker startup.
-type MCPServerConfig struct {
-	Command string            `mapstructure:"command" json:"command"`
-	Args    []string          `mapstructure:"args" json:"args,omitempty"`
-	Env     map[string]string `mapstructure:"env" json:"env,omitempty"`
-	URL     string            `mapstructure:"url" json:"url,omitempty"`
-}
-
-// Validate checks that the server config specifies exactly one of Command (stdio) or URL (remote).
-func (c *MCPServerConfig) Validate() error {
-	if c.Command == "" && c.URL == "" {
-		return fmt.Errorf("mcp server config: command or url required")
-	}
-	if c.Command != "" && c.URL != "" {
-		return fmt.Errorf("mcp server config: command and url are mutually exclusive")
-	}
-	return nil
-}
-
-// ClaudeCodeConfig holds Claude Code worker startup settings.
-type ClaudeCodeConfig struct {
-	Command               string                      `mapstructure:"command"`                 // binary + optional subcommand, e.g. "claude" or "ccr code"
-	PermissionPrompt      bool                        `mapstructure:"permission_prompt"`       // enable --permission-prompt-tool stdio for interaction chain
-	PermissionAutoApprove []string                    `mapstructure:"permission_auto_approve"` // tool names to auto-approve without user interaction
-	MCPServers            map[string]*MCPServerConfig `mapstructure:"mcp_servers"`             // user-configured MCP servers; empty = default discovery
-}
-
-// CodexCLIConfig holds Codex CLI worker startup settings.
-type CodexCLIConfig struct {
-	Command          string        `mapstructure:"command"`             // codex binary path, default "codex"
-	Model            string        `mapstructure:"model"`               // model name, empty = use Codex default
-	Sandbox          string        `mapstructure:"sandbox"`             // sandbox mode, default "danger-full-access" (YOLO: full filesystem + network)
-	ApprovalMode     string        `mapstructure:"approval_mode"`       // approval mode, default "never" (YOLO: no approval prompts)
-	Ephemeral        bool          `mapstructure:"ephemeral"`           // ephemeral sessions, default true
-	Personality      string        `mapstructure:"personality"`         // agent personality for app-server mode, default "friendly"
-	StartupTimeout   time.Duration `mapstructure:"startup_timeout"`     // process startup timeout, default 30s
-	CallTimeout      time.Duration `mapstructure:"call_timeout"`        // JSON-RPC call timeout, default 30s
-	UseAppServer     bool          `mapstructure:"use_app_server"`      // deprecated: always true, app-server is the only mode
-	IdleDrainPeriod  time.Duration `mapstructure:"idle_drain_period"`   // idle drain timeout for app-server mode, default 30m
-	Color            bool          `mapstructure:"color"`               // colored output (--color)
-	OutputFile       string        `mapstructure:"output_file"`         // output-only-last-message mode (--output-last-message)
-	StrictConfig     bool          `mapstructure:"strict_config"`       // strict config validation (--strict-config)
-	SkipGitRepoCheck bool          `mapstructure:"skip_git_repo_check"` // bypass git repo check (--skip-git-repo-check)
-	IgnoreUserConfig bool          `mapstructure:"ignore_user_config"`  // ignore user-level config (--ignore-user-config)
-	IgnoreRules      bool          `mapstructure:"ignore_rules"`        // ignore project rules (--ignore-rules)
-	LocalProvider    bool          `mapstructure:"local_provider"`      // force local model provider (--local-provider)
-	ConfigProfile    string        `mapstructure:"config_profile"`      // codex config profile (--profile)
-	BypassHookTrust  bool          `mapstructure:"bypass_hook_trust"`   // bypass hook trust (--dangerously-bypass-hook-trust)
-}
-
-// OpenCodeServerConfig holds OpenCode Server singleton process settings.
-type OpenCodeServerConfig struct {
-	Command           string        `mapstructure:"command"` // binary + optional subcommand, e.g. "opencode" or "opencode serve"
-	Password          string        `mapstructure:"password"`
-	IdleDrainPeriod   time.Duration `mapstructure:"idle_drain_period"`
-	ReadyTimeout      time.Duration `mapstructure:"ready_timeout"`
-	ReadyPollInterval time.Duration `mapstructure:"ready_poll_interval"`
-	HTTPTimeout       time.Duration `mapstructure:"http_timeout"`
-}
-
-// ACPConfig holds ACP (Agent Client Protocol) worker settings.
-// ACP is a universal worker type that connects to any ACP-compatible agent via stdio.
-type ACPConfig struct {
-	Command     string   `mapstructure:"command" json:"command"`                               // ACP agent binary (e.g. "hermes-acp")
-	AutoApprove *bool    `mapstructure:"auto_approve,omitempty" json:"auto_approve,omitempty"` // auto-approve permission requests
-	Args        []string `mapstructure:"args,omitempty" json:"args,omitempty"`                 // extra args appended after the command
-	Debug       bool     `mapstructure:"debug,omitempty" json:"debug,omitempty"`               // enable JSON-RPC protocol trace logging
-}
-
-// AutoRetryConfig controls automatic retry behavior when LLM provider returns
-// temporary errors (429 rate limit, 529 overload, 400 bad request, etc.).
-type AutoRetryConfig struct {
-	Enabled    bool          `mapstructure:"enabled"`
-	MaxRetries int           `mapstructure:"max_retries"`
-	BaseDelay  time.Duration `mapstructure:"base_delay"`
-	MaxDelay   time.Duration `mapstructure:"max_delay"`
-	RetryInput string        `mapstructure:"retry_input"`
-	NotifyUser bool          `mapstructure:"notify_user"`
-	Patterns   []string      `mapstructure:"patterns"`
-}
-
-// Defaults applies sensible defaults to AutoRetryConfig and returns the updated struct.
-func (c AutoRetryConfig) Defaults() AutoRetryConfig {
-	if c.MaxRetries <= 0 {
-		c.MaxRetries = 9
-	}
-	if c.BaseDelay <= 0 {
-		c.BaseDelay = 5 * time.Second
-	}
-	if c.MaxDelay <= 0 {
-		c.MaxDelay = 120 * time.Second
-	}
-	if c.RetryInput == "" {
-		c.RetryInput = "继续"
-	}
-	return c
-}
-
-// SecurityConfig holds auth and input validation settings.
-type SecurityConfig struct {
-	APIKeyHeader   string   `mapstructure:"api_key_header"`
-	APIKeys        []string `mapstructure:"api_keys"`
-	TLSEnabled     bool     `mapstructure:"tls_enabled"`
-	TLSCertFile    string   `mapstructure:"tls_cert_file"`
-	TLSKeyFile     string   `mapstructure:"tls_key_file"`
-	AllowedOrigins []string `mapstructure:"allowed_origins"`
-
-	// APIKeyUsers maps environment variable names (or literal key values) to user IDs.
-	// Enterprise multi-user: each API key gets a distinct identity for session isolation.
-	// Keys not present in this map default to "api_user" (backward compatible).
-	APIKeyUsers map[string]string `mapstructure:"api_key_users"`
-
-	// WorkDir security settings
-	WorkDirAllowedBasePatterns []string `mapstructure:"work_dir_allowed_base_patterns"` // extra whitelist patterns (supports ~ and ${VAR})
-	WorkDirForbiddenDirs       []string `mapstructure:"work_dir_forbidden_dirs"`        // extra blacklist directories
-
-	// CSP overrides the Content-Security-Policy header served with the embedded
-	// webchat SPA and docs portal. Empty string keeps the package-level default
-	// (localhost-friendly). Set this when serving from a remote host — the
-	// browser blocks fetch/ws/connect calls that do not match a directive in CSP.
-	CSP string `mapstructure:"csp"`
-
-	// SecurityContact enables /.well-known/security.txt (RFC 9116) when non-empty.
-	// Examples: "mailto:security@example.com", "https://example.com/security".
-	// Overridable via HOTPLEX_SECURITY_SECURITY_CONTACT env var.
-	SecurityContact string `mapstructure:"security_contact"`
-}
-
-// SessionConfig holds session lifecycle settings.
-type SessionConfig struct {
-	RetentionPeriod   time.Duration `mapstructure:"retention_period"`    // max session lifetime (default 7d)
-	GCScanInterval    time.Duration `mapstructure:"gc_scan_interval"`    // GC scan interval (default 1m)
-	MaxConcurrent     int           `mapstructure:"max_concurrent"`      // max concurrent sessions
-	TermRetention     time.Duration `mapstructure:"term_retention"`      // DB retention for terminated sessions (default 7d)
-	CronTermRetention time.Duration `mapstructure:"cron_term_retention"` // DB retention for terminated cron sessions (default 24h)
-}
-
-// PoolConfig holds session pool settings.
-type PoolConfig struct {
-	MinSize          int   `mapstructure:"min_size"`
-	MaxSize          int   `mapstructure:"max_size"`
-	MaxIdlePerUser   int   `mapstructure:"max_idle_per_user"`
-	MaxMemoryPerUser int64 `mapstructure:"max_memory_per_user"` // bytes; 0 = unlimited
-}
-
-type AgentConfig struct {
-	Enabled       bool     `mapstructure:"enabled"`        // enable agent config loading
-	ConfigDir     string   `mapstructure:"config_dir"`     // default: ~/.hotplex/agent-configs/
-	InjectExclude []string `mapstructure:"inject_exclude"` // global default: files to skip from agent config injection
-}
-
-// SkillsConfig holds skill discovery and caching settings.
-type SkillsConfig struct {
-	CacheTTL time.Duration `mapstructure:"cache_ttl"` // TTL for skills list cache, default 5m
-}
-
-// WebhookConfig holds GitHub webhook receiver settings.
-type WebhookConfig struct {
-	MaxBodySize   int64    `mapstructure:"max_body_size"`   // default: 1MB
-	AllowedRepos  []string `mapstructure:"allowed_repos"`   // repos to accept events from; empty = accept all (env override not supported for slices)
-	TargetJobName string   `mapstructure:"target_job_name"` // cron job to trigger on matching events
-	Secret        string   `mapstructure:"secret"`
-	Path          string   `mapstructure:"path"` // default: "/api/webhook/github"
-	Enabled       bool     `mapstructure:"enabled"`
-}
-
-// CronConfig holds AI-native cronjob scheduler settings.
-type CronConfig struct {
-	Enabled           bool             `mapstructure:"enabled"`
-	MaxConcurrentRuns int              `mapstructure:"max_concurrent_runs"` // default 3
-	MaxJobs           int              `mapstructure:"max_jobs"`            // default 50
-	DefaultTimeoutSec int              `mapstructure:"default_timeout_sec"` // default 300
-	TickIntervalSec   int              `mapstructure:"tick_interval_sec"`   // default 60
-	YAMLConfigPath    string           `mapstructure:"yaml_config_path"`    // optional external YAML
-	Jobs              []map[string]any `mapstructure:"jobs"`                // inline job definitions
-}
-
-// EventsConfig holds event and turn retention settings.
-type EventsConfig struct {
-	Retention time.Duration `mapstructure:"retention"` // TTL for events + turns, default 720h (30 days)
-}
-
-// ─── Defaults ────────────────────────────────────────────────────────────────
-
-// Default returns a Config with sensible production defaults.
-// All non-sensitive fields have values — the binary runs with zero config.
-func Default() *Config {
-	return &Config{
-		Gateway: GatewayConfig{
-			Addr:                  "localhost:8888",
-			ReadBufferSize:        4096,
-			WriteBufferSize:       4096,
-			PingInterval:          54 * time.Second,
-			PongTimeout:           60 * time.Second,
-			WriteTimeout:          10 * time.Second,
-			IdleTimeout:           5 * time.Minute,
-			MaxFrameSize:          32 * 1024,
-			BroadcastQueueSize:    256,
-			PlatformWriteBuffer:   64,
-			PlatformDropThreshold: 56,
-			DeltaCoalesceInterval: 120 * time.Millisecond,
-			DeltaCoalesceSize:     200,
-		},
-		DB: DBConfig{
-			Driver: "sqlite",
-			SQLite: SQLiteConfig{
-				Path:              filepath.Join(HotplexHome(), "data", "hotplex.db"),
-				WALMode:           true,
-				BusyTimeout:       5 * time.Second,
-				MaxOpenConns:      3, // 1 writer + 2 readers for shared session/event store
-				VacuumThreshold:   0.2,
-				CacheSizeKiB:      8192,
-				MmapSizeMiB:       64,
-				WalAutoCheckpoint: 2000,
-			},
-			// Legacy flat fields (kept for backward compat with existing consumers).
-			Path:              filepath.Join(HotplexHome(), "data", "hotplex.db"),
-			EventsPath:        "", // Deprecated: unused, events table lives in hotplex.db
-			WALMode:           true,
-			BusyTimeout:       5 * time.Second,
-			MaxOpenConns:      3, // 1 writer + 2 readers for shared session/event store
-			VacuumThreshold:   0.2,
-			CacheSizeKiB:      8192,
-			MmapSizeMiB:       64,
-			WalAutoCheckpoint: 2000,
-		},
-		Worker: WorkerConfig{
-			MaxLifetime:      24 * time.Hour,
-			IdleTimeout:      60 * time.Minute,
-			ExecutionTimeout: 30 * time.Minute,
-			TurnTimeout:      0, // disabled by default; execution_timeout catches zombies
-			EnvBlocklist:     nil,
-			DefaultWorkDir:   filepath.Join(HotplexHome(), "workspace"),
-			PIDDir:           filepath.Join(HotplexHome(), ".pids"),
-			AutoRetry:        AutoRetryConfig{Enabled: true, MaxRetries: 9, BaseDelay: 5 * time.Second, MaxDelay: 120 * time.Second, RetryInput: "继续", NotifyUser: true},
-			OpenCodeServer: OpenCodeServerConfig{
-				Command:           "opencode",
-				IdleDrainPeriod:   30 * time.Minute,
-				ReadyTimeout:      10 * time.Second,
-				ReadyPollInterval: 200 * time.Millisecond,
-				HTTPTimeout:       30 * time.Second,
-			},
-			ClaudeCode: ClaudeCodeConfig{
-				Command: "claude",
-			},
-			CodexCLI: CodexCLIConfig{
-				Command:         "codex",
-				Sandbox:         "danger-full-access",
-				ApprovalMode:    "never",
-				Ephemeral:       true,
-				Personality:     "friendly",
-				StartupTimeout:  30 * time.Second,
-				CallTimeout:     30 * time.Second,
-				UseAppServer:    true,
-				IdleDrainPeriod: 30 * time.Minute,
-			},
-		},
-		Security: SecurityConfig{
-			APIKeyHeader:    "X-API-Key",
-			APIKeys:         nil,
-			TLSEnabled:      false,
-			AllowedOrigins:  []string{"*"},
-			CSP:             "", // empty → webchat/docs use package-level default
-			SecurityContact: "",
-		},
-		Session: SessionConfig{
-			RetentionPeriod:   7 * 24 * time.Hour,
-			GCScanInterval:    1 * time.Minute,
-			MaxConcurrent:     1000,
-			TermRetention:     7 * 24 * time.Hour,
-			CronTermRetention: 24 * time.Hour,
-		},
-		Pool: PoolConfig{
-			MinSize:          0,
-			MaxSize:          100,
-			MaxIdlePerUser:   5,
-			MaxMemoryPerUser: 3 << 30, // 3 GB
-		},
-		Admin: AdminConfig{
-			Enabled:            true,
-			Addr:               "localhost:9999",
-			Tokens:             nil,
-			TokenScopes:        nil,
-			DefaultScopes:      []string{"session:read", "session:write", "session:delete", "stats:read", "health:read", "admin:write"},
-			IPWhitelistEnabled: false,
-			AllowedCIDRs:       []string{"127.0.0.0/8", "10.0.0.0/8"},
-			RateLimitEnabled:   true,
-			RequestsPerSec:     10,
-			Burst:              20,
-		},
-		Log: LogConfig{
-			Level:  "info",
-			Format: "json",
-		},
-		WebChat: WebChatConfig{
-			Addr:    "",
-			Enabled: true,
-		},
-		Messaging: MessagingConfig{
-			TurnSummaryEnabled: true,
-			// Shared defaults — propagated to platforms via propagateMessagingDefaults().
-			WorkerType: "claude_code",
-			STTConfig: STTConfig{
-				Provider:     "local",
-				LocalCmd:     "python3 " + filepath.Join(HotplexHome(), "scripts", "stt_server.py"),
-				LocalIdleTTL: time.Hour,
-			},
-			TTSConfig: TTSConfig{
-				TTSEnabled:      true,
-				TTSProvider:     "edge+moss",
-				Voice:           "zh-CN-XiaoxiaoNeural",
-				MaxChars:        150,
-				MossModelDir:    filepath.Join(HotplexHome(), "models", "moss-tts-nano"),
-				MossVoice:       "Xiaoyu",
-				MossPort:        18083,
-				MossIdleTimeout: 30 * time.Minute,
-				MossCpuThreads:  0,
-			},
-			Feishu: FeishuConfig{
-				MessagingPlatformConfig: defaultMessagingPlatformConfig(),
-			},
-			Slack: SlackConfig{
-				MessagingPlatformConfig: defaultMessagingPlatformConfig(),
-			},
-			Yuanxin: YuanxinConfig{
-				MessagingPlatformConfig: defaultMessagingPlatformConfig(),
-			},
-		},
-		AgentConfig: AgentConfig{
-			Enabled:   true,
-			ConfigDir: filepath.Join(HotplexHome(), "agent-configs"),
-		},
-		Skills: SkillsConfig{
-			CacheTTL: 5 * time.Minute,
-		},
-		Cron: CronConfig{
-			Enabled:           true,
-			MaxConcurrentRuns: 3,
-			MaxJobs:           50,
-			DefaultTimeoutSec: 300,
-			TickIntervalSec:   60,
-		},
-		Webhook: WebhookConfig{
-			MaxBodySize:   1 << 20, // 1MB
-			Path:          "/api/webhook/github",
-			TargetJobName: "pr-review-hotplex",
-			Enabled:       false,
-		},
-		Events: EventsConfig{
-			Retention: 720 * time.Hour, // 30 days
-		},
-	}
-}
-
-func defaultMessagingPlatformConfig() MessagingPlatformConfig {
-	return MessagingPlatformConfig{
-		RequireMention: true,
-		DMPolicy:       "allowlist",
-		GroupPolicy:    "allowlist",
-	}
-}
-
-// propagateMessagingDefaults copies shared fields from MessagingConfig to each
-// platform config. Only zero-value fields on the platform side are filled;
-// existing values are never overwritten.
-//
-// Priority: platform-level (YAML / env) > messaging-level > Default().
-func propagateMessagingDefaults(cfg *Config) {
-	msg := &cfg.Messaging
-	propagatePlatform(&msg.Slack.MessagingPlatformConfig, msg)
-	propagatePlatform(&msg.Feishu.MessagingPlatformConfig, msg)
-	propagatePlatform(&msg.Yuanxin.MessagingPlatformConfig, msg)
-
-	// Propagate shared defaults into per-bot configs.
-	for i := range msg.Slack.Bots {
-		propagateBotDefaults(&msg.Slack.MessagingPlatformConfig, &msg.Slack.Bots[i].STTConfig, &msg.Slack.Bots[i].TTSConfig)
-	}
-	for i := range msg.Feishu.Bots {
-		propagateBotDefaults(&msg.Feishu.MessagingPlatformConfig, &msg.Feishu.Bots[i].STTConfig, &msg.Feishu.Bots[i].TTSConfig)
-	}
-}
-
-// propagatePlatform fills zero-value fields on the platform from the messaging-level shared config.
-func propagatePlatform(p *MessagingPlatformConfig, msg *MessagingConfig) {
-	if p.WorkerType == "" {
-		p.WorkerType = msg.WorkerType
-	}
-	p.STTConfig.FillFrom(msg.STTConfig)
-	p.TTSConfig.FillFrom(msg.TTSConfig)
-}
-
-// normalizeSlackBots resolves SlackConfig to a unified Bots slice.
-// If Bots is already populated, it takes precedence.
-// If Bots is empty but top-level BotToken is set, auto-wraps as a single bot
-// with empty name (single-bot mode → agent-config uses platform-level directory).
-func normalizeSlackBots(cfg *SlackConfig) {
-	if len(cfg.Bots) > 0 {
-		return
-	}
-	if cfg.BotToken == "" {
-		return
-	}
-	cfg.Bots = []SlackBotConfig{
-		{Name: "", BotToken: cfg.BotToken, AppToken: cfg.AppToken},
-	}
-}
-
-// normalizeFeishuBots resolves FeishuConfig to a unified Bots slice.
-// Same backward-compat logic as normalizeSlackBots.
-func normalizeFeishuBots(cfg *FeishuConfig) {
-	if len(cfg.Bots) > 0 {
-		return
-	}
-	if cfg.AppID == "" {
-		return
-	}
-	cfg.Bots = []FeishuBotConfig{
-		{Name: "", AppID: cfg.AppID, AppSecret: cfg.AppSecret},
-	}
-}
-
-// propagateBotDefaults fills zero-value fields on each bot config from the
-// platform-level MessagingPlatformConfig and messaging-level shared config.
-func propagateBotDefaults(platformCfg *MessagingPlatformConfig, botSTT *STTConfig, botTTS *TTSConfig) {
-	botSTT.FillFrom(platformCfg.STTConfig)
-	botTTS.FillFrom(platformCfg.TTSConfig)
-}
-
-// expandStringFields expands env vars in non-empty string fields.
-func expandStringFields(fields ...*string) {
-	for _, f := range fields {
-		if *f != "" {
-			*f = ExpandEnv(*f)
-		}
-	}
-}
-
-// normalizePathFields resolves ~ and normalizes paths for non-empty string fields.
-func normalizePathFields(fields ...*string) {
-	for _, f := range fields {
-		if *f != "" {
-			if absPath, err := ExpandAndAbs(*f); err == nil {
-				*f = absPath
-			}
-		}
-	}
-}
-
-// ─── Loading ─────────────────────────────────────────────────────────────────
-
-// ErrConfigCycle is returned when a config inheritance chain contains a cycle.
-var ErrConfigCycle = errors.New("config: inheritance cycle detected")
-
-// Load reads configuration from the given file path, then applies defaults
-// and environment overrides.  Configuration strategy: convention over configuration.
-//
-// Load order (later overrides earlier):
-//  1. Sensible defaults (Default())
-//  2. Parent config file (via inherits field), recursively, with cycle detection
-//  3. Config file (YAML/JSON/TOML) — canonical source for non-sensitive values
-//  4. Environment variables (HOTPLEX_*)
-//
-// If filePath is empty, only defaults + environment are used.
-func Load(filePath string) (*Config, error) {
-	cfg, err := loadRecursive(filePath, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	// Environment variable overrides (e.g. HOTPLEX_LOG_FORMAT=text)
-	v := viper.New()
-	v.SetEnvPrefix("HOTPLEX")
-	v.SetEnvKeyReplacer(strings.NewReplacer(".", "_"))
-	v.AutomaticEnv()
-
-	// Explicitly bind common keys to ensure Unmarshal picks up environment variables.
-	// Viper's AutomaticEnv only works during Unmarshal if keys are known.
-	_ = v.BindEnv("log.level")
-	_ = v.BindEnv("log.format")
-	_ = v.BindEnv("db.path")
-	_ = v.BindEnv("db.wal_mode")
-	_ = v.BindEnv("db.driver")
-	_ = v.BindEnv("db.postgres.dsn")
-	_ = v.BindEnv("db.postgres.max_open_conns")
-	_ = v.BindEnv("gateway.addr")
-	_ = v.BindEnv("admin.enabled")
-	_ = v.BindEnv("admin.addr")
-	_ = v.BindEnv("session.max_concurrent")
-	_ = v.BindEnv("session.retention_period")
-	_ = v.BindEnv("session.term_retention")
-	_ = v.BindEnv("session.cron_term_retention")
-	_ = v.BindEnv("pool.max_size")
-	_ = v.BindEnv("pool.max_idle_per_user")
-	_ = v.BindEnv("pool.max_memory_per_user")
-	_ = v.BindEnv("worker.default_work_dir")
-	_ = v.BindEnv("worker.max_lifetime")
-	_ = v.BindEnv("worker.idle_timeout")
-	_ = v.BindEnv("worker.execution_timeout")
-	_ = v.BindEnv("worker.auto_retry.enabled")
-	_ = v.BindEnv("worker.auto_retry.max_retries")
-	_ = v.BindEnv("worker.claude_code.command")
-	_ = v.BindEnv("worker.codex_cli.command")
-	_ = v.BindEnv("worker.codex_cli.model")
-	_ = v.BindEnv("worker.codex_cli.sandbox")
-	_ = v.BindEnv("worker.codex_cli.approval_mode")
-	_ = v.BindEnv("worker.acp.command")
-	_ = v.BindEnv("worker.acp.auto_approve")
-	_ = v.BindEnv("worker.acp.args")
-	_ = v.BindEnv("worker.acp.debug")
-	_ = v.BindEnv("worker.opencode_server.command")
-	_ = v.BindEnv("worker.opencode_server.idle_drain_period")
-	_ = v.BindEnv("worker.opencode_server.ready_timeout")
-	_ = v.BindEnv("worker.opencode_server.ready_poll_interval")
-	_ = v.BindEnv("worker.opencode_server.http_timeout")
-	_ = v.BindEnv("worker.opencode_server.password")
-	_ = v.BindEnv("security.api_key_header")
-	_ = v.BindEnv("security.csp")
-	_ = v.BindEnv("security.allowed_origins")
-	_ = v.BindEnv("security.security_contact")
-	_ = v.BindEnv("agent_config.enabled")
-	_ = v.BindEnv("agent_config.config_dir")
-	_ = v.BindEnv("skills.cache_ttl")
-	_ = v.BindEnv("messaging.slack.work_dir")
-	_ = v.BindEnv("messaging.slack.stt_provider")
-	_ = v.BindEnv("messaging.slack.stt_local_cmd")
-	_ = v.BindEnv("messaging.slack.stt_local_idle_ttl")
-	_ = v.BindEnv("messaging.feishu.work_dir")
-	_ = v.BindEnv("messaging.feishu.stt_provider")
-	_ = v.BindEnv("messaging.feishu.stt_local_cmd")
-	_ = v.BindEnv("messaging.feishu.stt_local_idle_ttl")
-	_ = v.BindEnv("messaging.feishu.tts_enabled")
-	_ = v.BindEnv("messaging.feishu.tts_provider")
-	_ = v.BindEnv("messaging.feishu.tts_voice")
-	_ = v.BindEnv("messaging.feishu.tts_max_chars")
-	_ = v.BindEnv("messaging.feishu.tts_moss_model_dir")
-	_ = v.BindEnv("messaging.feishu.tts_moss_voice")
-	_ = v.BindEnv("messaging.feishu.tts_moss_port")
-	_ = v.BindEnv("messaging.feishu.tts_moss_idle_timeout")
-	_ = v.BindEnv("messaging.feishu.tts_moss_cpu_threads")
-	_ = v.BindEnv("messaging.slack.tts_enabled")
-	_ = v.BindEnv("messaging.slack.tts_provider")
-	_ = v.BindEnv("messaging.slack.tts_voice")
-	_ = v.BindEnv("messaging.slack.tts_max_chars")
-	_ = v.BindEnv("messaging.slack.tts_moss_model_dir")
-	_ = v.BindEnv("messaging.slack.tts_moss_voice")
-	_ = v.BindEnv("messaging.slack.tts_moss_port")
-	_ = v.BindEnv("messaging.slack.tts_moss_idle_timeout")
-	_ = v.BindEnv("messaging.slack.tts_moss_cpu_threads")
-	_ = v.BindEnv("webhook.enabled")
-	_ = v.BindEnv("webhook.secret")
-	_ = v.BindEnv("webhook.path")
-	_ = v.BindEnv("webhook.max_body_size")
-	_ = v.BindEnv("webhook.target_job_name")
-
-	if err := v.Unmarshal(cfg); err != nil {
-		return nil, fmt.Errorf("config: environment override: %w", err)
-	}
-
-	// Normalize path fields AFTER env overrides, because Viper's env binding
-	// writes raw values (e.g. "~/.hotplex/...") that bypass ExpandAndAbs.
-	cfg.normalizePaths()
-
-	return cfg, nil
-}
-
-// loadRecursive loads a config file and its ancestors, detecting cycles.
-// visited tracks file paths already loaded in the current chain; nil on the root call.
-func loadRecursive(filePath string, visited []string) (*Config, error) {
-	// Start with defaults.
-	cfg := Default()
-
-	// Track visited files for cycle detection.
-	var ancestors []string
-	if visited == nil {
-		ancestors = []string{}
-	} else {
-		ancestors = make([]string, len(visited), len(visited)+1)
-		copy(ancestors, visited)
-	}
-
-	var parentFile string
-	var childViper *viper.Viper
-
-	// If a config file is provided, unmarshal it over defaults.
-	if filePath != "" {
-		absPath, err := ExpandAndAbs(filePath)
-		if err != nil {
-			return nil, fmt.Errorf("config: resolve path %q: %w", filePath, err)
-		}
-		filePath = absPath
-
-		// Check for cycle: if this file is already in the ancestor chain.
-		if slices.Contains(ancestors, filePath) {
-			return nil, fmt.Errorf("%w: %v → %s", ErrConfigCycle, append(ancestors, filePath), filePath)
-		}
-
-		childViper = viper.New()
-		childViper.SetConfigFile(filePath)
-		if err := childViper.ReadInConfig(); err != nil {
-			return nil, fmt.Errorf("config: read %q: %w", filePath, err)
-		}
-		if err := childViper.Unmarshal(cfg); err != nil {
-			return nil, fmt.Errorf("config: unmarshal %q: %w", filePath, err)
-		}
-
-		parentFile = cfg.Inherits
-	}
-
-	// Recursively load parent config if inheritance is specified.
-	// Resolve parentFile relative to the directory of the current file.
-	if parentFile != "" {
-		ancestors = append(ancestors, filePath)
-		// If parentFile is relative, resolve it relative to the current file's directory.
-		if !filepath.IsAbs(parentFile) && filePath != "" {
-			parentFile = filepath.Join(filepath.Dir(filePath), parentFile)
-		}
-		parentCfg, err := loadRecursive(parentFile, ancestors)
-		if err != nil {
-			return nil, fmt.Errorf("config: inherits %q: %w", parentFile, err)
-		}
-		// Apply child values over parent using the already-loaded viper instance.
-		// This avoids a second disk read and eliminates TOCTOU risk.
-		if err := childViper.Unmarshal(parentCfg); err != nil {
-			return nil, fmt.Errorf("config: merge %q: %w", filePath, err)
-		}
-		*cfg = *parentCfg
-	}
-
-	// Expand env vars in token slices (supports ${VAR} references in config files).
-	for i, t := range cfg.Admin.Tokens {
-		cfg.Admin.Tokens[i] = ExpandEnv(t)
-	}
-	for i, k := range cfg.Security.APIKeys {
-		cfg.Security.APIKeys[i] = ExpandEnv(k)
-	}
-
-	// Numbered environment variables for slices (e.g. HOTPLEX_ADMIN_TOKEN_1..N)
-	// This supports project conventions for secret rotation and .env clarity.
-	cfg.Admin.Tokens = aggregateNumberedEnv(cfg.Admin.Tokens, "HOTPLEX_ADMIN_TOKEN_")
-	cfg.Security.APIKeys = aggregateNumberedEnv(cfg.Security.APIKeys, "HOTPLEX_SECURITY_API_KEY_")
-
-	// Resolve API key → user identity mappings from config.
-	cfg.ResolvedAPIKeyUsers = resolveAPIKeyUsers(cfg.Security.APIKeyUsers, cfg.Security.APIKeys)
-
-	// Messaging platform env var overrides.
-	applyMessagingEnv(cfg)
-
-	// Normalize multi-bot configs (backward compat: single-bot → bots[]).
-	normalizeSlackBots(&cfg.Messaging.Slack)
-	normalizeFeishuBots(&cfg.Messaging.Feishu)
-
-	// Propagate shared messaging defaults to per-platform configs.
-	// Priority: platform-level (YAML/env) > messaging-level > Default().
-	propagateMessagingDefaults(cfg)
-
-	// Expand env entries and drop those referencing unset vars without defaults.
-	expanded := make([]string, 0, len(cfg.Worker.Environment))
-	for _, e := range cfg.Worker.Environment {
-		if entry, ok := expandEnvEntry(e); ok {
-			expanded = append(expanded, entry)
-		} else if _, loaded := warnedEnvEntries.LoadOrStore(e, true); !loaded {
-			slog.Warn("excluding env entry: references unset variable", "entry", e)
-		}
-	}
-	cfg.Worker.Environment = expanded
-
-	cfg.normalizePaths()
-
-	return cfg, nil
-}
-
-// normalizePaths expands ~ and resolves relative paths for all config path fields,
-// and expands ${VAR} references in command templates.
-func (c *Config) normalizePaths() {
-	// 1. Expand environment variables in command templates and addresses.
-	for _, ef := range []*string{
-		&c.Gateway.Addr,
-		&c.Admin.Addr,
-		&c.Worker.ClaudeCode.Command,
-		&c.Worker.OpenCodeServer.Command,
-		&c.Worker.OpenCodeServer.Password,
-		&c.Worker.CodexCLI.Command,
-		&c.Messaging.Slack.LocalCmd,
-		&c.Messaging.Feishu.LocalCmd,
-		&c.Messaging.Feishu.MossModelDir,
-		&c.Messaging.Slack.MossModelDir,
-	} {
-		if *ef != "" {
-			*ef = ExpandEnv(*ef)
-		}
-	}
-	// Per-bot env expansion (credentials + paths).
-	var botFields []*string
-	for i := range c.Messaging.Slack.Bots {
-		botFields = append(botFields,
-			&c.Messaging.Slack.Bots[i].BotToken,
-			&c.Messaging.Slack.Bots[i].AppToken,
-			&c.Messaging.Slack.Bots[i].WorkDir,
-			&c.Messaging.Slack.Bots[i].LocalCmd,
-			&c.Messaging.Slack.Bots[i].MossModelDir,
-		)
-	}
-	for i := range c.Messaging.Feishu.Bots {
-		botFields = append(botFields,
-			&c.Messaging.Feishu.Bots[i].AppID,
-			&c.Messaging.Feishu.Bots[i].AppSecret,
-			&c.Messaging.Feishu.Bots[i].WorkDir,
-			&c.Messaging.Feishu.Bots[i].LocalCmd,
-			&c.Messaging.Feishu.Bots[i].MossModelDir,
-		)
-	}
-	expandStringFields(botFields...)
-
-	// 2. Expand ~ and normalize paths.
-	for _, pf := range []*string{
-		&c.DB.SQLite.Path,
-		&c.DB.Path,
-		&c.DB.EventsPath,
-		&c.Worker.DefaultWorkDir,
-		&c.Worker.PIDDir,
-		&c.AgentConfig.ConfigDir,
-		&c.Messaging.Slack.WorkDir,
-		&c.Messaging.Slack.MossModelDir,
-		&c.Messaging.Feishu.WorkDir,
-		&c.Messaging.Feishu.MossModelDir,
-		&c.Messaging.Yuanxin.WorkDir,
-	} {
-		if *pf != "" {
-			absPath, err := ExpandAndAbs(*pf)
-			if err != nil {
-				if _, loaded := warnedEnvEntries.LoadOrStore("path:"+*pf, true); !loaded {
-					slog.Warn("config: normalize path", "path", *pf, "err", err)
-				}
-				continue
-			}
-			*pf = absPath
-		}
-	}
-	// Per-bot WorkDir and MossModelDir path normalization.
-	var botPaths []*string
-	for i := range c.Messaging.Slack.Bots {
-		botPaths = append(botPaths, &c.Messaging.Slack.Bots[i].WorkDir, &c.Messaging.Slack.Bots[i].MossModelDir)
-	}
-	for i := range c.Messaging.Feishu.Bots {
-		botPaths = append(botPaths, &c.Messaging.Feishu.Bots[i].WorkDir, &c.Messaging.Feishu.Bots[i].MossModelDir)
-	}
-	normalizePathFields(botPaths...)
-}
-
-// ExpandAndAbs returns an absolute path, resolving ~ and relative paths.
-// If the path starts with ~ and $HOME is not set, the original path is returned.
-func ExpandAndAbs(p string) (string, error) {
-	if p == "" {
-		return "", nil
-	}
-	if strings.HasPrefix(p, "~/") {
-		home, err := os.UserHomeDir()
-		if err != nil {
-			// In test environments, $HOME may not be set. Return the path as-is
-			// rather than failing, but log a warning for visibility.
-			// The path will fail later when actually accessed, which is acceptable.
-			return p, nil
-		}
-		p = filepath.Join(home, p[2:])
-	}
-	if !filepath.IsAbs(p) {
-		abs, err := filepath.Abs(p)
-		if err != nil {
-			return "", err
-		}
-		p = abs
-	}
-	// Resolve symlinks to prevent TOCTOU attacks on SwitchWorkDir.
-	if resolved, err := filepath.EvalSymlinks(p); err == nil {
-		p = resolved
-	}
-	return p, nil
-}
-
-// ResolvePlatformWorkDir returns the workdir for the given platform,
-// falling back to Worker.DefaultWorkDir when the platform has no override.
-func (c *Config) ResolvePlatformWorkDir(platform string) string {
-	switch platform {
-	case "slack":
-		if c.Messaging.Slack.WorkDir != "" {
-			return c.Messaging.Slack.WorkDir
-		}
-	case "feishu":
-		if c.Messaging.Feishu.WorkDir != "" {
-			return c.Messaging.Feishu.WorkDir
-		}
-	case "yuanxin":
-		if c.Messaging.Yuanxin.WorkDir != "" {
-			return c.Messaging.Yuanxin.WorkDir
-		}
-	}
-	return c.Worker.DefaultWorkDir
-}
-
 // DefaultConfigPath is the default configuration file path used by the CLI
 // and the gateway. Defined here as the single source of truth to avoid
 // duplication across packages.
@@ -1382,265 +117,6 @@ func HotplexHome() string {
 		return TempBaseDir()
 	}
 	return filepath.Join(home, ".hotplex")
-}
-
-// aggregateNumberedEnv appends values from environment variables like PREFIX_1, PREFIX_2...
-// to the existing slice, deduplicating them.  Supports project's secret rotation convention.
-func aggregateNumberedEnv(existing []string, prefix string) []string {
-	seen := make(map[string]bool)
-	for _, v := range existing {
-		seen[v] = true
-	}
-
-	for i := 1; ; i++ {
-		key := fmt.Sprintf("%s%d", prefix, i)
-		val := os.Getenv(key)
-		if val == "" {
-			break
-		}
-		if !seen[val] {
-			existing = append(existing, val)
-			seen[val] = true
-		}
-	}
-	return existing
-}
-
-// resolveAPIKeyUsers builds a runtime map of expanded API key value → userID.
-// Input map keys are resolved as env var names first, then as literal keys.
-// Returns nil when no valid mappings are found (preserves "api_user" default).
-func resolveAPIKeyUsers(raw map[string]string, expandedKeys []string) map[string]string {
-	if len(raw) == 0 {
-		return nil
-	}
-	keySet := make(map[string]struct{}, len(expandedKeys))
-	for _, k := range expandedKeys {
-		keySet[k] = struct{}{}
-	}
-	result := make(map[string]string, len(raw))
-	for mapKey, userID := range raw {
-		if envVal := os.Getenv(mapKey); envVal != "" {
-			result[envVal] = userID
-			continue
-		}
-		if _, isKey := keySet[mapKey]; isKey {
-			result[mapKey] = userID
-		}
-	}
-	if len(result) == 0 {
-		return nil
-	}
-	return result
-}
-
-// applyMessagingEnv overrides messaging config from environment variables.
-// This is needed because Viper's AutomaticEnv cannot map nested keys
-// unless the viper instance has seen them from a config file or SetDefault.
-func applyMessagingEnv(cfg *Config) {
-	// Slack
-	applyPlatformEnv(&cfg.Messaging.Slack,
-		[]envMapping{
-			{"HOTPLEX_MESSAGING_SLACK_BOT_TOKEN", "BotToken"},
-			{"HOTPLEX_MESSAGING_SLACK_APP_TOKEN", "AppToken"},
-			{"HOTPLEX_MESSAGING_SLACK_WORKER_TYPE", "WorkerType"},
-			{"HOTPLEX_MESSAGING_SLACK_WORK_DIR", "WorkDir"},
-			{"HOTPLEX_MESSAGING_SLACK_DM_POLICY", "DMPolicy"},
-			{"HOTPLEX_MESSAGING_SLACK_GROUP_POLICY", "GroupPolicy"},
-		},
-		[]envMapping{
-			{"HOTPLEX_MESSAGING_SLACK_ENABLED", "Enabled"},
-			{"HOTPLEX_MESSAGING_SLACK_REQUIRE_MENTION", "RequireMention"},
-		},
-		[]envMapping{
-			{"HOTPLEX_MESSAGING_SLACK_ALLOW_FROM", "AllowFrom"},
-			{"HOTPLEX_MESSAGING_SLACK_ALLOW_DM_FROM", "AllowDMFrom"},
-			{"HOTPLEX_MESSAGING_SLACK_ALLOW_GROUP_FROM", "AllowGroupFrom"},
-		},
-	)
-
-	// Feishu
-	applyPlatformEnv(&cfg.Messaging.Feishu,
-		[]envMapping{
-			{"HOTPLEX_MESSAGING_FEISHU_APP_ID", "AppID"},
-			{"HOTPLEX_MESSAGING_FEISHU_APP_SECRET", "AppSecret"},
-			{"HOTPLEX_MESSAGING_FEISHU_WORKER_TYPE", "WorkerType"},
-			{"HOTPLEX_MESSAGING_FEISHU_WORK_DIR", "WorkDir"},
-			{"HOTPLEX_MESSAGING_FEISHU_DM_POLICY", "DMPolicy"},
-			{"HOTPLEX_MESSAGING_FEISHU_GROUP_POLICY", "GroupPolicy"},
-		},
-		[]envMapping{
-			{"HOTPLEX_MESSAGING_FEISHU_ENABLED", "Enabled"},
-			{"HOTPLEX_MESSAGING_FEISHU_REQUIRE_MENTION", "RequireMention"},
-		},
-		[]envMapping{
-			{"HOTPLEX_MESSAGING_FEISHU_ALLOW_FROM", "AllowFrom"},
-			{"HOTPLEX_MESSAGING_FEISHU_ALLOW_DM_FROM", "AllowDMFrom"},
-			{"HOTPLEX_MESSAGING_FEISHU_ALLOW_GROUP_FROM", "AllowGroupFrom"},
-		},
-	)
-
-	// Yuanxin
-	applyPlatformEnv(&cfg.Messaging.Yuanxin,
-		[]envMapping{
-			{"HOTPLEX_MESSAGING_YUANXIN_APP_ID", "AppID"},
-			{"HOTPLEX_MESSAGING_YUANXIN_PULSAR_URL", "PulsarURL"},
-			{"HOTPLEX_MESSAGING_YUANXIN_TENANT", "Tenant"},
-			{"HOTPLEX_MESSAGING_YUANXIN_NAMESPACE", "Namespace"},
-			{"HOTPLEX_MESSAGING_YUANXIN_PRODUCER_TOPIC", "ProducerTopic"},
-			{"HOTPLEX_MESSAGING_YUANXIN_WORKER_TYPE", "WorkerType"},
-			{"HOTPLEX_MESSAGING_YUANXIN_WORK_DIR", "WorkDir"},
-			{"HOTPLEX_MESSAGING_YUANXIN_DM_POLICY", "DMPolicy"},
-			{"HOTPLEX_MESSAGING_YUANXIN_GROUP_POLICY", "GroupPolicy"},
-		},
-		[]envMapping{
-			{"HOTPLEX_MESSAGING_YUANXIN_ENABLED", "Enabled"},
-		},
-		[]envMapping{
-			{"HOTPLEX_MESSAGING_YUANXIN_ALLOW_FROM", "AllowFrom"},
-			{"HOTPLEX_MESSAGING_YUANXIN_ALLOW_DM_FROM", "AllowDMFrom"},
-			{"HOTPLEX_MESSAGING_YUANXIN_ALLOW_GROUP_FROM", "AllowGroupFrom"},
-		},
-	)
-
-	// Global messaging env vars.
-	if v := os.Getenv("HOTPLEX_MESSAGING_TURN_SUMMARY_ENABLED"); v != "" {
-		cfg.Messaging.TurnSummaryEnabled = strings.EqualFold(v, "true")
-	}
-
-	// Messaging-level shared defaults (propagated to platforms by propagateMessagingDefaults).
-	msgStrs := []envMapping{
-		{"HOTPLEX_MESSAGING_WORKER_TYPE", "WorkerType"},
-		{"HOTPLEX_MESSAGING_STT_PROVIDER", "Provider"},
-		{"HOTPLEX_MESSAGING_STT_LOCAL_CMD", "LocalCmd"},
-		{"HOTPLEX_MESSAGING_TTS_PROVIDER", "TTSProvider"},
-		{"HOTPLEX_MESSAGING_TTS_VOICE", "Voice"},
-		{"HOTPLEX_MESSAGING_TTS_MOSS_MODEL_DIR", "MossModelDir"},
-		{"HOTPLEX_MESSAGING_TTS_MOSS_VOICE", "MossVoice"},
-	}
-	for _, m := range msgStrs {
-		if v := os.Getenv(m.env); v != "" {
-			if err := setField(&cfg.Messaging, m.field, v); err != nil {
-				slog.Warn("config: env mapping skipped",
-					"env", m.env,
-					"field", m.field,
-					"target", "config.Messaging",
-					"error", err,
-				)
-			}
-		}
-	}
-	// Int fields.
-	if v := os.Getenv("HOTPLEX_MESSAGING_TTS_MAX_CHARS"); v != "" {
-		if n, err := strconv.Atoi(v); err == nil {
-			cfg.Messaging.MaxChars = n
-		}
-	}
-	if v := os.Getenv("HOTPLEX_MESSAGING_TTS_MOSS_PORT"); v != "" {
-		if n, err := strconv.Atoi(v); err == nil {
-			cfg.Messaging.MossPort = n
-		}
-	}
-	if v := os.Getenv("HOTPLEX_MESSAGING_TTS_MOSS_CPU_THREADS"); v != "" {
-		if n, err := strconv.Atoi(v); err == nil {
-			cfg.Messaging.MossCpuThreads = n
-		}
-	}
-	// Duration fields.
-	if v := os.Getenv("HOTPLEX_MESSAGING_STT_LOCAL_IDLE_TTL"); v != "" {
-		if d, err := time.ParseDuration(v); err == nil {
-			cfg.Messaging.LocalIdleTTL = d
-		}
-	}
-	if v := os.Getenv("HOTPLEX_MESSAGING_TTS_MOSS_IDLE_TIMEOUT"); v != "" {
-		if d, err := time.ParseDuration(v); err == nil {
-			cfg.Messaging.MossIdleTimeout = d
-		}
-	}
-	// Bool fields.
-	if v := os.Getenv("HOTPLEX_MESSAGING_TTS_ENABLED"); v != "" {
-		cfg.Messaging.TTSEnabled = strings.EqualFold(v, "true")
-	}
-}
-
-// envMapping maps an environment variable to a struct field name.
-type envMapping struct{ env, field string }
-
-// applyPlatformEnv applies string, bool, and slice env-var mappings to a target struct.
-func applyPlatformEnv(target any, strs, bools, slices []envMapping) {
-	for _, m := range strs {
-		if v := os.Getenv(m.env); v != "" {
-			if err := setField(target, m.field, v); err != nil {
-				slog.Warn("config: env mapping skipped",
-					"env", m.env,
-					"field", m.field,
-					"target", fmt.Sprintf("%T", target),
-					"error", err,
-				)
-			}
-		}
-	}
-	for _, m := range bools {
-		if v := os.Getenv(m.env); v != "" {
-			if err := setBoolField(target, m.field, strings.EqualFold(v, "true")); err != nil {
-				slog.Warn("config: env mapping skipped",
-					"env", m.env,
-					"field", m.field,
-					"target", fmt.Sprintf("%T", target),
-					"error", err,
-				)
-			}
-		}
-	}
-	for _, m := range slices {
-		if v := os.Getenv(m.env); v != "" {
-			if err := setSliceField(target, m.field, v); err != nil {
-				slog.Warn("config: env mapping skipped",
-					"env", m.env,
-					"field", m.field,
-					"target", fmt.Sprintf("%T", target),
-					"error", err,
-				)
-			}
-		}
-	}
-}
-
-// setField sets a string field on a struct by name using reflection.
-func setField(target any, field, value string) error {
-	v := reflect.ValueOf(target).Elem()
-	f := v.FieldByName(field)
-	if !f.IsValid() {
-		return fmt.Errorf("config: setField: no such field %q on %T", field, target)
-	}
-	f.SetString(value)
-	return nil
-}
-
-// setBoolField sets a bool field on a struct by name using reflection.
-func setBoolField(target any, field string, value bool) error {
-	v := reflect.ValueOf(target).Elem()
-	f := v.FieldByName(field)
-	if !f.IsValid() {
-		return fmt.Errorf("config: setBoolField: no such field %q on %T", field, target)
-	}
-	f.SetBool(value)
-	return nil
-}
-
-// setSliceField sets a []string field on a struct by name using reflection.
-func setSliceField(target any, field, value string) error {
-	v := reflect.ValueOf(target).Elem()
-	f := v.FieldByName(field)
-	if !f.IsValid() {
-		return fmt.Errorf("config: setSliceField: no such field %q on %T", field, target)
-	}
-	parts := strings.Split(value, ",")
-	slice := make([]string, len(parts))
-	for i, p := range parts {
-		slice[i] = strings.TrimSpace(p)
-	}
-	f.Set(reflect.ValueOf(slice))
-	return nil
 }
 
 // ResolveInjectExclude resolves the inject_exclude list using 3-level fallback:
